+++ app/src/main/java/de/ironjan/mensaupb/stw/filters/Filter.java
@@ -9,10 +9,18 @@ import de.ironjan.mensaupb.stw.*;
  */
 interface Filter {
     /**
-     * Filters the given menus
+     * Filters the given menus. No guarantee on order is given
      *
      * @param menus the menus to be filtered
      * @return filtered clone
      */
     public List<RawMenu> filter(List<RawMenu> menus);
+
+    /**
+     * Filters the given menus
+     *
+     * @param menu the menu to be filtered
+     * @return filtered clone
+     */
+    public RawMenu filter(RawMenu menu);
 }
